+++ src/main/java/com/example/modeladov1/service/ResenaService.java
@@ -4,11 +4,14 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
 
+import com.example.modeladov1.model.EstadoPedido;
 import com.example.modeladov1.model.Resena;
 import com.example.modeladov1.repository.ResenaRepository;
 import com.example.modeladov1.model.Producto;
 import com.example.modeladov1.model.Usuario;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -22,44 +25,48 @@ public class ResenaService {
     @Autowired
     ProductoService productoService; // Inyecta productoService
 
-    public List<Resena> getAll(){
+    public ResponseEntity<List<Resena>> getAll(){
         List<Resena> resenas = new ArrayList<>();
         for(Resena resena : repo.findAll()){
             resenas.add(resena);
         }
-        return resenas;
+        return new ResponseEntity<>(resenas, HttpStatus.OK);
     }
 
-    public Resena getOne(Integer id) {
-        return repo.findById(id).orElse(null);
+    public ResponseEntity<Resena> getOne(Integer id) {
+        Resena objeto = repo.findById(id).orElse(null);
+        if(objeto!=null){
+            return new ResponseEntity<>(objeto, HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public void add(Resena resena){
-        Usuario usuario = usuarioService.getOne(resena.getUsuario().getId_usuario()); // Usa usuarioService para llamar a getOne
-        Producto producto = productoService.getProductoById(resena.getProducto().getId_producto()); // Usa productoService para llamar a getOne
-        resena.setUsuario(usuario); // Establece el Usuario en la Resena
-        resena.setProducto(producto); // Establece el Producto en la Resena
-        repo.save(resena);
+    public ResponseEntity<Resena> add(Resena resena){
+        Resena objeto = repo.save(resena);
+        return new ResponseEntity<>(objeto, HttpStatus.OK);
     }
 
-    public void eliminarResena(int id) {
-        repo.deleteById(id);
+    public ResponseEntity<Resena> eliminarResena(int id) {
+        if(repo.findById(id).isPresent()){
+            repo.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }else{
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
-    public Resena actualizarResena(Integer id_resena, Resena resenaActualizada) {
+    public ResponseEntity<Resena> actualizarResena(Integer id_resena, Resena resenaActualizada) {
         Optional<Resena> resenaExistente = repo.findById(id_resena);
 
         if (resenaExistente.isPresent()) {
             Resena resena = resenaExistente.get();
-            Usuario usuario = usuarioService.getOne(resenaActualizada.getUsuario().getId_usuario()); // Usa usuarioService para llamar a getOne
-            Producto producto = productoService.getProductoById(resenaActualizada.getProducto().getId_producto()); // Usa productoService para llamar a getOne
-            resena.setUsuario(usuario); // Establece el Usuario en la Resena
-            resena.setProducto(producto); // Establece el Producto en la Resena
             resena.setResena(resenaActualizada.getResena());
             resena.setCalificacion(resenaActualizada.getCalificacion());
-            return repo.save(resena);
+            Resena objeto = repo.save(resena);
+            return new ResponseEntity<>(objeto,HttpStatus.OK);
         } else {
-            throw new NoSuchElementException("Reseña no encontrada");
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         }
     }
 }
